Changed static_modifiers.txt so that all difficulty levels are affected in varying degrees. Added a localisation of normal_difficulty
so that it shows as "Normal Difficulty" in-game instead of
normal_difficulty, similar to how all the other difficulties behaved.
However, there is still one bug regarding normal difficulty static
modifiers. In the options, it shows as "No effect" even though the
effects are in place in-game.

Also removed monthly_character_wealth as it seemingly had no effect.